refactor(time-picker): new appearance
