+++ src/time-picker/panel/TimePickerRangePanel.tsx
@@ -7,7 +7,7 @@ import SinglePanel, { SinglePanelProps } from './SinglePanel';
 import useConfig from '../../_util/useConfig';
 import Button from '../../button';
 
-import { DEFAULT_STEPS, DEFAULT_FORMAT, useTimePickerTextConfig } from '../consts';
+import { DEFAULT_STEPS, DEFAULT_FORMAT, useTimePickerTextConfig } from '../const';
 
 import { TdTimeRangePickerProps, TimeRangeValue } from '../type';
